weather_scraping/requirements.txt
Django==5.0.4
psycopg2-binary==2.9.9
djangorestframework==3.15.1
setuptools==69.5.1
django-redis==5.4.0
redis==5.0.4
python-dotenv==1.0.0
pytest-django==4.5.2
pytest==7.2.0
django-filter==24.2
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.6.0
